services/identity-service/requirements.txt
Django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
psycopg[binary]>=3.1
gunicorn>=21.2
alembic
django-prometheus
prometheus-client
pytest>=8.0
pytest-django>=4.8